refactor(visualization): relocate feature-private utils into their slices
Move utils that are used exclusively by a single feature out of util/ and
co-locate them next to their consumer, following the feature-slice
architecture (features have no util/ folder):

- selectTopNByValue -> features/codeMap/
- formatCompactNumber -> features/sidebarExplorer/
- clipboardWriter -> features/viewCubeToolbox/services/
- loadFilesValidationToErrorDialog -> features/loadFile/services/
- uploadFiles/{createCCFileInput,readFiles} -> features/navBar/services/
- pipes/removeExtension.pipe -> features/navBar/

Also relocate the misplaced loadFilesValidationToErrorDialog.spec.ts (it
sat under util/uploadFiles/) next to its source, and fix the clipboardWriter
spec filename casing. Shared utils (algorithm/, codeMapHelper, color/,
blacklist/, ...) remain in util/.
