+++ visualization/app/codeCharta/features/loadFile/services/loadFile.service.spec.ts
@@ -11,7 +11,7 @@ import packageJson from "../../../../../package.json"
 import { clone } from "../../../util/clone"
 import { klona } from "klona"
 import { ErrorDialogService } from "../../../features/shared/components/errorDialog/errorDialog.service"
-import { loadFilesValidationToErrorDialog } from "../../../util/loadFilesValidationToErrorDialog"
+import { loadFilesValidationToErrorDialog } from "./loadFilesValidationToErrorDialog"
 import { fileRoot } from "../../../util/fileRoot"
 import { metricDataSelector } from "../../../state/selectors/accumulatedData/metricData/metricData.selector"
 import { State, Store, StoreModule } from "@ngrx/store"
