+++ visualization/app/codeCharta/features/loadFile/services/loadFile.service.ts
@@ -5,7 +5,7 @@ import { CCFileValidationResult } from "../../../util/fileValidator"
 import { FileState } from "../../../model/files/files"
 import { NameDataPair } from "../../../codeCharta.model"
 import { ErrorDialogService } from "../../../features/shared/components/errorDialog/errorDialog.service"
-import { loadFilesValidationToErrorDialog } from "../../../util/loadFilesValidationToErrorDialog"
+import { loadFilesValidationToErrorDialog } from "./loadFilesValidationToErrorDialog"
 import { enrichFileStatesAndRecentFilesWithValidationResults } from "../../../util/fileParser"
 import { fileRoot } from "../../../util/fileRoot"
 import { LoadFileStore } from "../stores/loadFile.store"
