+++ visualization/app/codeCharta/features/loadFile/services/loadInitialFile.service.ts
@@ -9,7 +9,7 @@ import { CcState, NameDataPair } from "../../../codeCharta.model"
 import { FileState } from "../../../model/files/files"
 import { MetricQueryParemter } from "../../../state/effects/updateQueryParameters/metricQueryParameter"
 import { ErrorDialogService } from "../../../features/shared/components/errorDialog/errorDialog.service"
-import { buildHtmlMessage } from "../../../util/loadFilesValidationToErrorDialog"
+import { buildHtmlMessage } from "./loadFilesValidationToErrorDialog"
 import { getNameDataPair } from "../../../util/fileParser"
 import { LoadFileService, NO_FILES_LOADED_ERROR_MESSAGE } from "./loadFile.service"
 import { UrlExtractor } from "../../../util/urlExtractor"
