+++ visualization/app/codeCharta/features/navBar/components/mapSelector/mapSelector.component.ts
@@ -2,7 +2,7 @@ import { ChangeDetectionStrategy, Component, computed, ElementRef, HostListener,
 import { toSignal } from "@angular/core/rxjs-interop"
 import { CCFile } from "../../../../codeCharta.model"
 import { FileSelectionState, FileState } from "../../../../model/files/files"
-import { RemoveExtensionPipe } from "../../../../util/pipes/removeExtension.pipe"
+import { RemoveExtensionPipe } from "../../removeExtension.pipe"
 import { FilesSelectionStore } from "../../stores/filesSelection.store"
 
 type FileRemovedInUIState = {
