+++ visualization/app/codeCharta/features/navBar/removeExtension.pipe.spec.ts
