+++ visualization/app/codeCharta/features/navBar/services/readFiles.spec.ts
