+++ visualization/app/codeCharta/features/sidebarExplorer/formatCompactNumber.spec.ts
