+++ visualization/app/codeCharta/features/sidebarExplorer/formatCompactNumber.ts
