+++ visualization/app/codeCharta/features/viewCubeToolbox/services/clipboardWriter.spec.ts
