+++ visualization/app/codeCharta/features/viewCubeToolbox/services/clipboardWriter.ts
